feat: 댓글 삭제 시 비밀번호 비교 기능 추가
- body의 password와 DB의 password 비교 후 일치 시 삭제
- 불일치 시 401 삭제 권한 없음 메시지 반환

+++ routes/comments.js
@@ -83,7 +83,7 @@ router.put("/posts/:_postId/comments/:_commentId", async (req, res) => {
                 await Comments.updateOne({ commentId: _commentId }, { $set: { content, password } })
                 // 201, 수정 완료 메시지 반환
                 return res.status(201).json({ message: "댓글을 수정하였습니다." })
-                // 비밀번호 불일치 시, 401 삭제권한 없음 메시지 반환
+                // 비밀번호 불일치 시, 401 수정권한 없음 메시지 반환
             } else {
                 return res.status(401).json({ message: "수정 권한이 없습니다." })
             }
@@ -112,10 +112,16 @@ router.delete("/posts/:_postId/comments/:_commentId", async (req, res) => {
             .where("commentId").equals(_commentId)
         // 검색결과가 있는 경우,
         if (result.length) {
-            // DB에서 삭제 후,
-            await Comments.deleteOne({ commentId: _commentId })
-            // 200, 삭제 완료 메시지 반환
-            return res.status(200).json({ message: "댓글을 삭제하였습니다." })
+            // 비밀번호 일치 시
+            if (password === result[0].password) {
+                // DB에서 삭제 후,
+                await Comments.deleteOne({ commentId: _commentId })
+                // 200, 삭제 완료 메시지 반환
+                return res.status(200).json({ message: "댓글을 삭제하였습니다." })
+                // 비밀번호 불일치 시, 401 삭제권한 없음 메시지 반환
+            } else {
+                return res.status(401).json({ message: "삭제 권한이 없습니다." })
+            }
             // 검색 결과가 없는 경우 400, 게시글 조회 실패 메시지 반환
         } else {
             return res.status(404).json({ message: "댓글 조회에 실패하였습니다." })
